requirements.txt
requests==2.20.1
Scrapy==1.5.1
Pillow==5.4.1
matplotlib==3.0.3
pandas==0.24.2
pytesseract==0.2.6
